pyproject.toml
[tool.poetry]
name = "testovoeexel"
version = "0.1.0"
description = ""
authors = ["griega <126513423+griega@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.8.1"
openpyxl = "^3.0.9"
python-dotenv = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
